remove unused parsing code
